- Split sources
- Slightly changed the textures of the tab button
