+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/ShadowHunter22sConfigLibraryClient.java
@@ -7,15 +7,11 @@ package dev.shadowhunter22.shadowhunter22sconfiglibrary;
 
 import net.fabricmc.api.ClientModInitializer;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 public class ShadowHunter22sConfigLibraryClient implements ClientModInitializer {
-	public static final String MOD_ID = "shadowhunter22s-config-library";
-	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
+	public static final String MOD_ID = ShadowHunter22sConfigLibrary.MOD_ID;
 
 	@Override
 	public void onInitializeClient() {
-		LOGGER.info("Successfully loaded {}!", MOD_ID);
+		ShadowHunter22sConfigLibrary.LOGGER.info("Successfully loaded {}!", MOD_ID);
 	}
 }
