+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/annotation/ConfigEntry.java
