+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/api/v1/config/AutoConfigManager.java
