+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/api/v1/gui/widget/category/CategoryTab.java
