+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/option/type/EnumConfigOption.java
