+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/serializer/AbstractSerializer.java
