+++ src/client/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/util/TranslationUtil.java
