+++ src/client/resources/assets/shadowhunter22s-config-library/textures/gui/header_separator.png
