+++ src/client/resources/assets/shadowhunter22s-config-library/textures/gui/reset_button.png
