+++ src/client/resources/shadowhunter22s-config-library.client.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "dev.shadowhunter22.shadowhunter22sconfiglibrary.mixin.client",
+  "package": "dev.shadowhunter22.shadowhunter22sconfiglibrary.mixin",
   "compatibilityLevel": "JAVA_17",
   "client": [
 	"SliderWidgetInvoker",
