+++ src/main/java/dev/shadowhunter22/shadowhunter22sconfiglibrary/ShadowHunter22sConfigLibrary.java
@@ -1,0 +1,16 @@
+package dev.shadowhunter22.shadowhunter22sconfiglibrary;
+
+import net.fabricmc.api.ModInitializer;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ShadowHunter22sConfigLibrary implements ModInitializer {
+	public static final String MOD_ID = "shadowhunter22s-config-library";
+	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
+
+	@Override
+	public void onInitialize() {
+		// client-only
+	}
+}
