+++ src/test/resources/assets/shadowhunter22s-config-library-testmod/lang/en_us.json
@@ -1,20 +1,45 @@
 {
+  "text.config-library.TEST_2.@Category": "Category 1",
+  "text.config-library.TEST_4.@Category": "Category 2",
+
+  "screen.title.config-library": "Config Library Test Config",
+
+  "text.config-library.TEST_2.@Section": "Test 2 Section",
+  "option.config-library.TEST_2": "Test 2",
+
+  "option.config-library.TEST_3": "Test 3",
+
+  "text.config-library.TEST_4.@Section": "Test 4 Section",
+  "option.config-library.TEST_4": "Test 4",
+
+  "option.config-library.TEST_5": "Test 5",
   "option.config-library.TEST_5.TopLeft": "Top Left",
   "option.config-library.TEST_5.TopRight": "Top Right",
   "option.config-library.TEST_5.BottomLeft": "Bottom Left",
   "option.config-library.TEST_5.BottomRight": "Bottom Right",
+
+  "text.config-library.TEST_6.@Section": "Test 6 Section",
+  "option.config-library.TEST_6": "Test 6",
   "option.config-library.TEST_6.TopLeft": "Top Left",
   "option.config-library.TEST_6.TopRight": "Top Right",
   "option.config-library.TEST_6.BottomLeft": "Bottom Left",
   "option.config-library.TEST_6.BottomRight": "Bottom Right",
+
+  "option.config-library.TEST_7": "Test 7",
   "option.config-library.TEST_7.TopLeft": "Top Left",
   "option.config-library.TEST_7.TopRight": "Top Right",
   "option.config-library.TEST_7.BottomLeft": "Bottom Left",
   "option.config-library.TEST_7.BottomRight": "Bottom Right",
+
+  "option.config-library.TEST_8": "Test 8",
   "option.config-library.TEST_8.TopLeft": "Top Left",
   "option.config-library.TEST_8.TopRight": "Top Right",
   "option.config-library.TEST_8.BottomLeft": "Bottom Left",
   "option.config-library.TEST_8.BottomRight": "Bottom Right",
-  "text.config-library.TEST_2.@Category": "Category 1",
-  "text.config-library.TEST_4.@Category": "Category 2"
+
+  "option.config-library.TEST_9": "Test 9",
+  "option.config-library.TEST_9.TopLeft": "Top Left",
+  "option.config-library.TEST_9.TopRight": "Top Right",
+  "option.config-library.TEST_9.BottomLeft": "Bottom Left",
+  "option.config-library.TEST_9.BottomRight": "Bottom Right"
 }
